% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LADprobs.R
\name{LADprobs}
\alias{LADprobs}
\title{LADprobs
2016 Sean Thomas}
\usage{
LADprobs(binFile, countFile, ladFile, viralChr)
}
\arguments{
\item{binFile}{a bin file in HiTC format}

\item{countFile}{a counts file in HiTC format}

\item{ladFile}{a 2 column file, first column comntains all host genome bin IDs from bin file, second column is an indicator(0/1) showing whether or not each bin overlaps a LAD.}

\item{viralChr}{a viral genome name: 'chrebv' for example}
}
\value{
vector of LAD state prediction probabilities for all bins in bin file
}
\description{
LADprobs
2016 Sean Thomas
}
\examples{
ladStateProbs <- LADprobs(binFile,countFile,ladFile,viralChr)
}
